Finished working proof of concept

+++ src/client/java/io/github/mikip98/ofg/property/Util.java
@@ -4,5 +4,5 @@ public class Util {
     public static short autoFloodFillMinValue = 512;
     public static short autoFloodFillMaxValue = 8191;
 
-    public static int floodFillIgnoreFirstEntryId = 50;
+    public static short floodFillIgnoreFirstEntryId = 50;
 }
